Refactor DOM geometry modules into dom/geometry submodule
