+++ components/script/dom/bindings/structuredclone.rs
@@ -46,10 +46,10 @@ use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::serializable::{Serializable, StorageKey};
 use crate::dom::bindings::transferable::Transferable;
 use crate::dom::blob::Blob;
-use crate::dom::dompoint::DOMPoint;
-use crate::dom::dompointreadonly::DOMPointReadOnly;
 use crate::dom::file::File;
 use crate::dom::filelist::FileList;
+use crate::dom::geometry::dompoint::DOMPoint;
+use crate::dom::geometry::dompointreadonly::DOMPointReadOnly;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::imagebitmap::ImageBitmap;
 use crate::dom::imagedata::ImageData;
