+++ components/script/dom/geometry/mod.rs
@@ -1,0 +1,8 @@
+pub(crate) mod dommatrix;
+pub(crate) mod dommatrixreadonly;
+pub(crate) mod dompoint;
+pub(crate) mod dompointreadonly;
+pub(crate) mod domquad;
+pub(crate) mod domrect;
+pub(crate) mod domrectlist;
+pub(crate) mod domrectreadonly;
