+++ components/script/dom/mod.rs
@@ -254,16 +254,11 @@ pub(crate) mod dissimilaroriginwindow;
 pub(crate) mod document;
 pub(crate) use self::document::*;
 pub(crate) mod domexception;
+
+pub(crate) mod geometry;
+pub(crate) use self::geometry::*;
 pub(crate) mod domimplementation;
-pub(crate) mod dommatrix;
-pub(crate) mod dommatrixreadonly;
 pub(crate) mod domparser;
-pub(crate) mod dompoint;
-pub(crate) mod dompointreadonly;
-pub(crate) mod domquad;
-pub(crate) mod domrect;
-pub(crate) mod domrectlist;
-pub(crate) mod domrectreadonly;
 pub(crate) mod domstringlist;
 pub(crate) mod domstringmap;
 pub(crate) mod domtokenlist;
