+++ components/script/dom/range.rs
@@ -34,9 +34,9 @@ use crate::dom::bindings::weakref::{WeakRef, WeakRefVec};
 use crate::dom::characterdata::CharacterData;
 use crate::dom::document::Document;
 use crate::dom::documentfragment::DocumentFragment;
-use crate::dom::domrect::DOMRect;
-use crate::dom::domrectlist::DOMRectList;
 use crate::dom::element::Element;
+use crate::dom::geometry::domrect::DOMRect;
+use crate::dom::geometry::domrectlist::DOMRectList;
 use crate::dom::html::htmlscriptelement::HTMLScriptElement;
 use crate::dom::node::{Node, NodeTraits, ShadowIncluding};
 use crate::dom::selection::Selection;
