+++ components/script/dom/webxr/xrpose.rs
@@ -7,7 +7,7 @@ use dom_struct::dom_struct;
 use crate::dom::bindings::codegen::Bindings::XRPoseBinding::XRPoseMethods;
 use crate::dom::bindings::reflector::{Reflector, reflect_dom_object};
 use crate::dom::bindings::root::{Dom, DomRoot};
-use crate::dom::dompointreadonly::DOMPointReadOnly;
+use crate::dom::geometry::dompointreadonly::DOMPointReadOnly;
 use crate::dom::window::Window;
 use crate::dom::xrrigidtransform::XRRigidTransform;
 use crate::dom::xrsession::ApiRigidTransform;
