+++ components/script/webdriver_handlers.rs
@@ -70,9 +70,9 @@ use crate::dom::bindings::reflector::{DomGlobal, DomObject};
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::DOMString;
 use crate::dom::document::Document;
-use crate::dom::domrect::DOMRect;
 use crate::dom::element::Element;
 use crate::dom::eventtarget::EventTarget;
+use crate::dom::geometry::domrect::DOMRect;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::html::htmlbodyelement::HTMLBodyElement;
 use crate::dom::html::htmldatalistelement::HTMLDataListElement;
